requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
black==24.3.0
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
diskcache==5.6.3
distro==1.9.0
    # via openai
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via openai
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
mypy-extensions==1.0.0
    # via black
openai==1.14.1
packaging==24.0
    # via black
pathspec==0.12.1
    # via black
platformdirs==4.2.0
    # via black
pydantic==2.6.4
    # via openai
pydantic-core==2.16.3
    # via pydantic
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tqdm==4.66.2
    # via openai
typing-extensions==4.10.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
